Create function for detecting filter changes on URL
